mkbootabledisk: allow boards to customize everything via vendor scripts
With more and more boards to support which differ in various ways w.r.t.
the way in which they want their firmware image programmed, certain
files appended to the boot partition and/or rootfs, things don't scale
so well if we keep adding stuff as command-line options to
mkbootabledisk, I'll eventually forget how to supply those arguments.

Replace that system with a vendor script system which overrides some
bash functions and variables defined in the central script.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>